[ag] #131 - refacto integration (WIP) [tests] - conversion::as_tuple
